requirements.txt
agno
python-dotenv
streamlit 

duckduckgo-search
beautifulsoup4

groq
sentence-transformers
chromadb

langchain_unstructured
unstructured[all-docs]
langchain_community